src/DeltaShell.Plugins.FMSuite.FlowFM/IO/CrossSectionDefinitionGenerators/CrossSectionDefinitionIniCategoryGeneratorElliptical.cs
using System;
using DelftTools.Hydro.CrossSections;
using DelftTools.Hydro.CrossSections.StandardShapes;
using DeltaShell.NGHS.IO.FileWriters.CrossSectionDefinition;
using DeltaShell.NGHS.IO.Helpers;

namespace DeltaShell.Plugins.FMSuite.FlowFM.IO.CrossSectionDefinitionGenerators
{
    public class CrossSectionDefinitionIniCategoryGeneratorElliptical : ACrossSectionDefinitionIniCategoryGenerator
    {
        public override DelftIniCategory GenerateIniCategory(CrossSectionDefinitionStandard crossSectionDefinition)
        {
            var crossSectionShape = crossSectionDefinition.Shape as CrossSectionStandardShapeElliptical;
            if (crossSectionShape == null) throw new Exception();

            return base.GenerateIniCategory(crossSectionDefinition);
        }

        protected override void AddMeasurementsProperties(ICrossSectionStandardShape crossSectionShape)
        {
            var ellipticalShape = crossSectionShape as CrossSectionStandardShapeElliptical;
            iniCategory.AddProperty(DefinitionPropertySettings.EllipseWidth.Key, $"{ellipticalShape.Width:0.00}");
            iniCategory.AddProperty(DefinitionPropertySettings.EllipseHeight.Key, $"{ellipticalShape.Height:0.00}");
        }
    }
}
